requirements.txt
# python==3.10.14
# just for the app.py
pandas==2.1.3
streamlit==1.40.2
Pillow==9.3.0